Major code refactor
Huge changes to support the newest opengl/glsl api. Screw opengl for making
some of the things so deprecated...but it's for the best now. Various
changes include the shader syntax and functionality, and the introduction of
completely user-defined matrices. Much more versatile, I approve of it now,
but that was a pain to transition. Must fix perspective issues soon and data
is NOT getting passed correctly to the fragment shader currently.
Loads models perfectly fine, and has basic lighting/shader capability. Will
continue to build upon this!

Signed-off-by: Tudor Marcu <tudormarcu4@gmail.com>